run-command-on-git-revisions aborts if run on a dirty repo
It would not be nice to run the run-command-on-git-revisions on a dirty repo. I added a check with some hints from [stackoverflow](http://stackoverflow.com/a/2659808). The script now aborts if you have staged or unstaged changes or even untracked files.

Maybe I could add a --force flag to run anyway?